refactor: export parseContainmentPath for use in app
