+++ packages/common/src/core/index.ts
@@ -14,6 +14,7 @@ export * from "./shareEntityWithGroups";
 export * from "./unshareEntityWithGroups";
 export * from "./getEntityGroups";
 export * from "./getEntityThumbnailUrl";
+export * from "./parseContainmentPath";
 
 // For sme reason, if these are exported here,
 // they are not actually exported in the final package.
